+++ crates/cubecl-convolution/src/homogeneous/simple_tma.rs
@@ -55,7 +55,7 @@ where
         >,
 {
     type LhsLoader = TmaIm2colLoader<MP, Self::Config>;
-    type Config = ConvolutionConfig<single_stage::Config<SMM::Config>>;
+    type Config = ConvolutionConfig<single_stage::SingleStageConfig<SMM::Config>>;
     type RhsLoader = TmaWeightLoader<MP, SMM::Config>;
     type AccumulatorLoader = BiasLoader<MP>;
 
@@ -189,7 +189,7 @@ impl<SMM> ConvolutionConfigFactory for SimpleTmaConvolutionFamily<SMM>
 where
     SMM: StageMatmulFamily,
 {
-    type Config = config::ConvolutionConfig<single_stage::Config<SMM::Config>>;
+    type Config = config::ConvolutionConfig<single_stage::SingleStageConfig<SMM::Config>>;
     type Input = GlobalInput<SMM::Input>;
 
     fn check_config(config: &Self::Config) -> Result<(), InvalidConfigError> {
@@ -222,7 +222,7 @@ where
         let stage_k = stage_config.tiling_scheme().elements_in_stage_k();
 
         config::ConvolutionConfig::new(
-            single_stage::Config::new(
+            single_stage::SingleStageConfig::new(
                 stage_config,
                 // TODO: Find the correct condition to avoid check bounds.
                 true,
