+++ crates/cubecl-matmul/src/components/global/base.rs
@@ -3,8 +3,8 @@ use cubecl_core::prelude::*;
 
 use crate::{
     components::{
-        Ident, InputIdent, InvalidConfigError, LoadSpecializationConfig, MatmulLineSizes,
-        MatmulPrecision, MatmulProblem, MatrixLayout, TilingScheme,
+        Ident, InputIdent, InvalidConfigError, LoadSpecializationConfig, MatmulChecker,
+        MatmulLineSizes, MatmulPrecision, MatmulProblem, MatrixLayout, TilingScheme,
         config::MatmulConfig,
         global::{
             PlaneRoleConfig, RoleRuleConfig, SpecializedLoadingSides, multi_stage::EventLoadingMode,
@@ -21,9 +21,8 @@ use cubecl_std::{
 use super::{GlobalWriter, Quantization, load::LoaderMode};
 
 /// A family of [matmuls](GlobalMatmul) working with any [precision](MatmulPrecision).
-pub trait GlobalMatmulFamily: Send + Sync + 'static {
+pub trait GlobalMatmulFamily: Send + Sync + 'static + MatmulChecker<Config: GlobalConfig> {
     type Matmul<MP: MatmulPrecision>: GlobalMatmul<MP, Config = Self::Config>;
-    type Config: GlobalConfig;
     type Input;
 
     fn cube_dim(
@@ -36,7 +35,6 @@ pub trait GlobalMatmulFamily: Send + Sync + 'static {
         problem: &MatmulProblem,
         line_sizes: &MatmulLineSizes,
         cube_dim: &CubeDim,
-        cube_count: &CubeCount,
         quantized: bool,
     ) -> Self::Config;
 }
