+++ crates/cubecl-matmul/src/components/global/multi_stage/double_buffer_execution.rs
