+++ crates/cubecl-matmul/src/components/global/multi_stage/double_buffering/mod.rs
@@ -1,5 +1,6 @@
 mod config;
 mod matmul;
+mod setup;
 
 pub use config::*;
 pub use matmul::*;
