+++ crates/cubecl-matmul/src/components/global/multi_stage/mod.rs
@@ -1,7 +1,7 @@
 pub mod double_buffering;
 pub mod ordered;
 
+mod double_buffer_execution;
 mod event_listener;
-mod execute;
 
 pub use event_listener::*;
