+++ crates/cubecl-matmul/src/components/global/multi_stage/ordered/mod.rs
@@ -1,5 +1,13 @@
 mod config;
 mod matmul;
+mod setup;
 
 pub use config::*;
 pub use matmul::*;
+
+use crate::components::global::load::sync_full_ordered;
+
+/// The ordered double buffering global matmul
+/// requires tilewise loading on `Lhs` to guarantee that planes
+/// only use data they have loaded themselves.
+pub type LL = sync_full_ordered::LoadingStrategy;
