+++ crates/cubecl-matmul/src/components/global/single_stage/barrier/mod.rs
