+++ crates/cubecl-matmul/src/components/global/single_stage/mod.rs
@@ -1,4 +1,6 @@
+pub mod barrier;
 pub mod simple;
+pub mod tma;
 
 mod config;
